Arrglo bug de que cuadno se hace un update a un usuario no se borre la foto si no es lo q se quiere, ambien implemento la opcion de borrar un usuario

+++ src/EventListener/FotoEquipoUploadListener.php
@@ -46,8 +46,11 @@ class FotoEquipoUploadListener
     public function preUpdate(PreUpdateEventArgs $args)
     {
         $entity = $args->getEntity();
-
+        if( $entity instanceof Usuario && empty($entity->getFotoFile())){
+            return;
+        }
         if($entity instanceof Equipo || $entity instanceof Usuario){
+
             if($entity->getFoto()!==null && $entity->getFoto()!==''){
 
                 unlink(__DIR__.'/../../public/uploads/fotos/'.$entity->getFoto());
